Refactor package hierachy in core.trust
