+++ examples/Trust/models/wine_quality_model.py
@@ -2,16 +2,13 @@ import h1st as h1
 import sys
 
 
-from h1st.core.trust.explainers.shap_explainer import SHAPExplainer
-from h1st.core.trust.explainers.lime_explainer import LIMEExplainer
-
 import os
 import boto3
 import pandas as pd
 from sklearn.model_selection import train_test_split
 from sklearn.ensemble import RandomForestRegressor
 from sklearn.metrics import accuracy_score, f1_score
-import matplotlib.pyplot as plt
+#import matplotlib.pyplot as plt
 import sklearn
 
 
@@ -92,24 +89,3 @@ class WineQualityModel(h1.Model):
         df["quality"] = df["quality"].astype(int)
         input_data = df[self.features]
         return self.model.predict(input_data)
-
-    def describe(self, constituent=h1.Model.Constituency.DATA_SCIENTIST.name, aspect=h1.Model.Aspect.ACCOUNTABLE.name):
-        ## TODO: For each pair of constituent and aspect write functions to show relevant information.
-        print("Description type: {}"\
-            .format(h1.Model.Aspect.ACCOUNTABLE.name))
-        print("Targeted Constituent: {}"\
-            .format(h1.Model.Constituency.DATA_SCIENTIST.name))
-        print("Model Metrics : {}".format(self.metrics))
-        
-        print("Size of the WineQuality dataset: {}".format(self.data.shape[0]))
-        print("Number of features of the WineQuality dataset: {}".format(len(self.features)))
-
-        if self.shap:
-            print("Overview of the features that are most important for the WineQualityModel. Seen in the plot are SHAP values of every feature for every sample.The plot below sorts features by the sum of SHAP value magnitudes over all samples, and uses SHAP values to show the distribution of the impacts each feature has on the model output. The color represents the feature value (red high, blue low). This reveals for example that a high alcohol increases the predicted wine quality.".format())
-            d = SHAPExplainer(self.model, self.prepared_data, self.plot)      
-            return {'shap_values':d.shap_values}
-
-    def explain(self, constituent=h1.Model.Constituency.DATA_SCIENTIST.name, aspect=h1.Model.Aspect.ACCOUNTABLE.name, decision=None):
-        e = LIMEExplainer(self.model, self.prepared_data, decision, self.plot)
-        return {'lime_predictions':e.decision_explainer.as_list()}
-
