--- h1st/core/trust/explainers/utils.py
@@ -1,8 +1,0 @@
-def get_model_name(model):
-    return str(type(model).__name__)
-
-def get_constituent():
-    pass
-
-def get_aspect():
-    pass
