+++ h1st/core/trust/shap_model_describer.py
@@ -1,11 +1,8 @@
-import pandas as pd
-import shap
 import numpy as np
-import matplotlib.pyplot as plt
+import shap
 
 
-
-class SHAPExplainer():
+class SHAPModelDescriber():
     def __init__(self, model, data, plot):
         super().__init__()
         self.model = model
@@ -16,23 +13,21 @@ class SHAPExplainer():
         self.plot = plot
         if plot:
             self.generate_plots()
-        
-    
-    
+
     def return_shap_values(self):
         return {'shap_values':self.shap_values}
- 
+
     def _shap_local_plot(self, j):
         explainer_model = shap.TreeExplainer(self.model)
         shap_values_model = explainer_model.shap_values(self.samples)
-        
+
         p = shap.force_plot(
             explainer_model.expected_value,
             shap_values_model[j],
             self.samples.iloc[[j]],
         )
         return p
- 
+
     def _random_sample_generator(self):
         val_df = self.data["val_df"].copy()
         val_df.loc[:, "predict"] = np.round(self.model.predict(val_df), 2)
@@ -53,4 +48,3 @@ class SHAPExplainer():
         df = self.data["train_df"].reset_index(drop=True)
         shap.summary_plot(self.shap_values, self.data["train_df"],)       
         # shap.force_plot(self.shap_explainer.expected_value, self.shap_values, df)
-
